steganography-app/steganography-app/requirements.txt
PyQt5>=5.15.0
PyQt5-tools>=5.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
cryptography>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.5.0
flask>=2.0.0
requests>=2.25.0
psutil>=5.8.0
pywavelets>=1.1.0
numba>=0.56.0
pyqtgraph>=0.11.0
pytest>=6.2.4